+++ src/AsteroidShapeModels.jl
@@ -69,7 +69,9 @@ export load_shape_obj, load_shape_grid, grid_to_faces
 export polyhedron_volume, equivalent_radius, maximum_radius, minimum_radius
 
 include("visibility.jl")
-export view_factor, build_face_visibility_graph!, isilluminated
+export view_factor, build_face_visibility_graph!
+export isilluminated, update_illumination!
+export apply_eclipse_shadowing!, EclipseStatus, NO_ECLIPSE, PARTIAL_ECLIPSE, TOTAL_ECLIPSE
 
 include("geometry_utils.jl")
 export angle_rad, angle_deg, solar_phase_angle, solar_elongation_angle